# Replace with flexible CSV parser