This change is intended to make it more extensible than existing parser.
It is aimed at ease of addition of options and improvement of maintainability of code.

Benchmark

```
# before change

Warming up --------------------------------------
            unquoted     5.000  i/100ms
              quoted     2.000  i/100ms
     include col_sep     1.000  i/100ms
     include row_sep     1.000  i/100ms
        encode utf-8     4.000  i/100ms
         encode sjis     4.000  i/100ms
Calculating -------------------------------------
            unquoted     59.692  (± 3.4%) i/s -    300.000  in   5.029315s
              quoted     27.762  (± 3.6%) i/s -    140.000  in   5.049609s
     include col_sep     15.317  (± 6.5%) i/s -     77.000  in   5.039037s
     include row_sep      6.526  (±15.3%) i/s -     33.000  in   5.130936s
        encode utf-8     46.373  (± 6.5%) i/s -    232.000  in   5.039316s
         encode sjis     46.922  (± 4.3%) i/s -    236.000  in   5.038045s
ruby benchmark/parse.rb  42.17s user 0.25s system 99% cpu 42.795 total

# after change

Warming up --------------------------------------
            unquoted     4.000  i/100ms
              quoted     1.000  i/100ms
     include col_sep     1.000  i/100ms
     include row_sep     1.000  i/100ms
        encode utf-8     3.000  i/100ms
         encode sjis     3.000  i/100ms
Calculating -------------------------------------
            unquoted     45.147  (± 4.4%) i/s -    228.000  in   5.057924s
              quoted     18.500  (± 5.4%) i/s -     93.000  in   5.033393s
     include col_sep     18.845  (± 5.3%) i/s -     95.000  in   5.045361s
     include row_sep      7.360  (± 0.0%) i/s -     37.000  in   5.032146s
        encode utf-8     37.648  (± 5.3%) i/s -    189.000  in   5.029352s
         encode sjis     35.967  (± 2.8%) i/s -    180.000  in   5.011789s
```